Rename ThroughputTracker to UsageMonitor and add some docstrings
